requirements.txt
streamlit
google-generativeai
pymupdf
python-docx
python-pptx
markdown2
pdfkit
graphviz